award page and refactored some stuff
