+++ src/App.js
@@ -1,11 +1,11 @@
 import React from 'react';
 import { BrowserRouter as Router, Route } from 'react-router-dom';
 import './styles/styles.scss';
-import Navbar from './navbar/navbar';
-import Research from './research/research';
-import About from './about/about';
-import Footer from './navbar/footer';
-
+import Navbar from './components/navbar/navbar';
+import Research from './components/research/research';
+import About from './components/about/about';
+import Awards from './components/awards/awards';
+import Footer from './components/navbar/footer';
 
 function App() {
   return (
@@ -15,6 +15,7 @@ function App() {
           <div className="body-part">
             <About />
             <Research />
+            <Awards />
           </div>
           <Footer />
       </div>
