+++ src/components/research/research.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import researchProjects from '../data/research_projects.json';
+import researchProjects from '../../data/research_projects.json';
 import ResearchPreview from './research_preview.js';
 
 const Research = () =>
@@ -9,11 +9,11 @@ const Research = () =>
         Publications
       </div>
       <div>
-        {researchProjects.map((research) => <ResearchPreview id={research.id} research={research}/>)}
+        {researchProjects.map((research) =>
+          <ResearchPreview key={research.id} research={research}
+          />)}
       </div>
     </div>
   );
-
-
 
 export default Research;
