+++ src/data/awards.json
@@ -1,0 +1,14 @@
+[
+  {
+    "id": 1,
+    "title":"EECS Best SuperUROP Award",
+    "description": "Awarded by the MIT EECS department for performing well in undergraduate research."
+  },
+  {
+    "id": 2,
+    "title": "Leiserchess Performance Engineering Award",
+    "description": "Awarded by the class staff of 6.172 Performance Engineering of Software Systems for designing and implementing the fastest chess-bot in class."
+  }
+
+
+]
