+++ src/styles/general.scss
@@ -20,3 +20,7 @@ $desktop: "screen and (min-width : 1281px)";
   border-bottom: 1px solid lightgrey;
 }
 
+.body-part {
+  margin: 60px 50px 20px 50px;
+}
+
