+++ src/styles/navbar.scss
@@ -7,7 +7,7 @@ nav {
   width: 100%;
   height: 30px;
 
-  background: gray;
+  background: darkgrey;
   display: flex;
   align-items: flex-end;
   font-family: $titleFont;
