+++ src/styles/research.scss
@@ -19,7 +19,6 @@
         img {
           width: 200px;
         }
-
       }
 
       .descriptions {
