+++ src/styles/styles.scss
@@ -2,6 +2,7 @@
 @import './navbar.scss';
 @import './about.scss';
 @import './research.scss';
+@import './awards.scss';
 
 div.page-container {
   margin: 0;
@@ -11,6 +12,3 @@ body {
   margin: 0;
 }
 
-.body-part {
-  margin: 60px 20px 20px 20px;
-}
